Delete GCIM_HazardCurveApp and merge GCIM features into main HazardCurveApplication
